pyproject.toml
[tool.poetry]
name = "tantra-massage-bot"
version = "0.1.0"
description = ""
authors = ["zai <artziga77@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
